[oto.comoco-light-dark]: fixed issue with history hightchart block, added light/dark theme (cherry-pick from develop)
